server/pyproject.toml
[project]
name = "langchain-demo"
version = "0.1.0"
description = "AI apps using LangChain"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "jq>=1.8.0",
    "langchain>=0.1.20",
    "langchain-community>=0.0.38",
    "langchain-openai>=0.1.7",
    "langchain-qdrant>=0.1.4",
    "numpy>=1.26.4",
    "python-dotenv>=1.0.1",
    "qdrant-client[fastembed]>=1.12.1",
]
